Mouse movements output moves in standard notation.
Changed the logic of mouseDown and mouseUp functions, they now output moves in standard notation instead of just moving the cube. This will be useful when sending moves to the the server etc.

This was an attempt to make the logic simpler. Both functions still need further refactoring and commenting to make the code more readable.